Refactor main donation form to separate component
